requirements.txt
asgiref==3.3.1
astroid==2.5
autopep8==1.5.5
Django==3.1.7
djangorestframework==3.12.2
flake8==3.8.4
importlib-metadata==3.5.0
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
wrapt==1.12.1
zipp==3.4.0
